Put moved files into the correct package.
